A lot of structural refactoring and package renaming
