+++ leanpub/src/main/groovy/org/asciidoctor/converter/LeanpubConverterRegistry.groovy
@@ -1,12 +1,12 @@
-package org.asciidoctor.leanpub.internal
+package org.asciidoctor.converter
 
 import org.asciidoctor.Asciidoctor
-import org.asciidoctor.leanpub.LeanpubConverter
+import org.asciidoctor.converter.LeanpubConverter
 
 /**
  * @author Schalk W. Cronjé
  */
-class ConverterRegistry implements org.asciidoctor.converter.spi.ConverterRegistry {
+class LeanpubConverterRegistry implements org.asciidoctor.converter.spi.ConverterRegistry {
     @Override
     void register(Asciidoctor asciidoctor) {
         asciidoctor.javaConverterRegistry().register(LeanpubConverter,'leanpub')
