+++ leanpub/src/main/resources/META-INF/services/org.asciidoctor.converter.spi.ConverterRegistry
@@ -1,1 +1,1 @@
-org.asciidoctor.leanpub.internal.ConverterRegistry
+org.asciidoctor.converter.LeanpubConverterRegistry
