+++ markdown/src/main/resources/META-INF/services/org.asciidoctor.converter.spi.ConverterRegistry
@@ -1,1 +1,1 @@
-org.asciidoctor.markdown.internal.ConverterRegistry
+org.asciidoctor.converter.MarkdownConverterRegistry
